Split attribute mutations
