+++ saleor/graphql/attribute/mutations/__init__.py
@@ -1,0 +1,23 @@
+from .attribute_create import AttributeCreate
+from .attribute_delete import AttributeDelete
+from .attribute_reorder_values import AttributeReorderValues
+from .attribute_update import AttributeUpdate
+from .attribute_value_create import AttributeValueCreate
+from .attribute_value_delete import AttributeValueDelete
+from .attribute_value_update import AttributeValueUpdate
+from .base_reorder_attributes import (
+    BaseReorderAttributesMutation,
+    BaseReorderAttributeValuesMutation,
+)
+
+__all__ = [
+    "AttributeCreate",
+    "AttributeDelete",
+    "AttributeReorderValues",
+    "AttributeUpdate",
+    "AttributeValueCreate",
+    "AttributeValueDelete",
+    "AttributeValueUpdate",
+    "BaseReorderAttributesMutation",
+    "BaseReorderAttributeValuesMutation",
+]
